src/requirements.txt
chainlit==0.7.700
cohere==4.37
openai==1.3.5
tiktoken==0.5.1
python-dotenv==1.0.0xz